For some reason uint16_t is about 10% faster than uint8_t

+++ Experiments/Workloads/Dijkstra/Dijkstra-S/main.c
@@ -9,7 +9,7 @@
 #include <stdlib.h>
 #include <limits.h>
 
-typedef uint8_t dtype;
+typedef uint16_t dtype;
 
 typedef struct {
 	dtype vertex;
